refactor(test): 패키지 변경, BeforEach 적용
* day, order 패키지 추가
* DiscountEventManagerTest: @BeforeEach 추가
